+++ src/main/java/org/geowe/client/local/layermanager/toolbar/RasterLayerToolBar.java
@@ -29,7 +29,7 @@ import javax.inject.Inject;
 
 import org.geowe.client.local.layermanager.tool.DeleteLayerTool;
 import org.geowe.client.local.layermanager.tool.InfoRasterTool;
-import org.geowe.client.local.layermanager.tool.create.AddWMSLayerTool;
+import org.geowe.client.local.layermanager.tool.create.AddRasterLayerTool;
 
 import com.google.gwt.user.client.ui.HorizontalPanel;
 import com.sencha.gxt.core.client.resources.ThemeStyles;
@@ -43,7 +43,7 @@ public class RasterLayerToolBar extends ContentPanel {
 	@Inject @New
 	private InfoRasterTool infoRasterTool;
 	@Inject
-	private AddWMSLayerTool addWMSLayerTool;
+	private AddRasterLayerTool addWMSLayerTool;
 	private final HorizontalPanel horizontalGroup;
 	
 	public RasterLayerToolBar(){