+++ CHANGELOG.md
@@ -7,8 +7,22 @@ You can see the previous changelog [here] (http://www.geowe.org/index.php?id=cha
 ### Added
 - More spatial tools
 
+## [1.1.0]Beta - 2016-08-31
+
+### Added
+- Added WMTS PNOA layer into Catalog
+- Added WMTS and TMS support
+- Added zoom level info panel
+
+### Fixed
+- issue [#151](https://github.com/geowe/geowe-core/issues/151)
+- issue [#152](https://github.com/geowe/geowe-core/issues/152)
+- issue [#153](https://github.com/geowe/geowe-core/issues/153)
+- issue [#159](https://github.com/geowe/geowe-core/issues/159)
+
 ## [1.0.5]Beta - 2016-08-19
 version deployed for testing [map.geowe.org](http://map.geowe.org)
+
 ### Added
 - Added Africa projections: EPSG:22032, EPSG:22033, EPSG:32732, EPSG:3036, EPSG:3037, EPSG:32737
 
@@ -18,6 +32,7 @@ version deployed for testing [map.geowe.org](http://map.geowe.org)
 
 ## [1.0.4]Beta - 2016-07-22
 version deployed for testing [map.geowe.org](http://map.geowe.org) no longer available
+
 ### Fixed
 - issue [#147](https://github.com/geowe/geowe-core/issues/147)
 - issue [#148](https://github.com/geowe/geowe-core/issues/147)